feat: 파일 이름 변경 및 FollowList,FollowingList 랜더링 기능 추가
